+++ packages/flutter_driver/lib/src/common/message.dart
@@ -4,6 +4,10 @@
 
 import 'package:meta/meta.dart';
 
+import 'error.dart';
+import 'find.dart';
+import 'wait.dart';
+
 /// An object sent from the Flutter Driver to a Flutter application to instruct
 /// the application to perform a task.
 abstract class Command {
@@ -66,4 +70,164 @@ abstract class Result {
 
   /// Serializes this message to a JSON map.
   Map<String, dynamic> toJson();
+}
+
+/// A Flutter Driver command aimed at an object to be located by [finder].
+///
+/// Implementations must provide a concrete [kind]. If additional data is
+/// required beyond the [finder] the implementation may override [serialize]
+/// and add more keys to the returned map.
+abstract class CommandWithTarget extends Command {
+  /// Constructs this command given a [finder].
+  CommandWithTarget(this.finder, {Duration timeout}) : super(timeout: timeout) {
+    if (finder == null)
+      throw DriverError('$runtimeType target cannot be null');
+  }
+
+  /// Deserializes this command from the value generated by [serialize].
+  CommandWithTarget.deserialize(Map<String, String> json)
+      : finder = SerializableFinder.deserialize(json),
+        super.deserialize(json);
+
+  /// Locates the object or objects targeted by this command.
+  final SerializableFinder finder;
+
+  /// This method is meant to be overridden if data in addition to [finder]
+  /// is serialized to JSON.
+  ///
+  /// Example:
+  ///
+  ///     Map<String, String> toJson() => super.toJson()..addAll({
+  ///       'foo': this.foo,
+  ///     });
+  @override
+  Map<String, String> serialize() =>
+      super.serialize()..addAll(finder.serialize());
+}
+
+/// A Flutter Driver command that waits until [finder] can locate the target.
+class WaitFor extends CommandWithTarget {
+  /// Creates a command that waits for the widget identified by [finder] to
+  /// appear within the [timeout] amount of time.
+  ///
+  /// If [timeout] is not specified the command times out after 5 seconds.
+  WaitFor(SerializableFinder finder, {Duration timeout})
+      : super(finder, timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitFor.deserialize(Map<String, String> json) : super.deserialize(json);
+
+  @override
+  String get kind => 'waitFor';
+}
+
+/// The result of a [WaitFor] command.
+class WaitForResult extends Result {
+  /// Creates a [WaitForResult].
+  const WaitForResult();
+
+  /// Deserializes the result from JSON.
+  static WaitForResult fromJson(Map<String, dynamic> json) {
+    return const WaitForResult();
+  }
+
+  @override
+  Map<String, dynamic> toJson() => <String, dynamic>{};
+}
+
+/// A Flutter Driver command that waits until [finder] can no longer locate the target.
+class WaitForAbsent extends CommandWithTarget {
+  /// Creates a command that waits for the widget identified by [finder] to
+  /// disappear within the [timeout] amount of time.
+  ///
+  /// If [timeout] is not specified the command times out after 5 seconds.
+  WaitForAbsent(SerializableFinder finder, {Duration timeout})
+      : super(finder, timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitForAbsent.deserialize(Map<String, String> json) : super.deserialize(json);
+
+  @override
+  String get kind => 'waitForAbsent';
+}
+
+/// The result of a [WaitForAbsent] command.
+class WaitForAbsentResult extends Result {
+  /// Creates a [WaitForAbsentResult].
+  const WaitForAbsentResult();
+
+  /// Deserializes the result from JSON.
+  static WaitForAbsentResult fromJson(Map<String, dynamic> json) {
+    return const WaitForAbsentResult();
+  }
+
+  @override
+  Map<String, dynamic> toJson() => <String, dynamic>{};
+}
+
+/// A Flutter Driver command that waits until a given [condition] is satisfied.
+class WaitForCondition extends Command {
+  /// Creates a command that waits for the given [condition] is met.
+  const WaitForCondition(this.condition, {Duration timeout}) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitForCondition.deserialize(Map<String, String> json)
+      : condition = WaitConditionDecoder.deserialize(json),
+        super.deserialize(json);
+
+  /// The condition that this command shall wait for.
+  final WaitCondition condition;
+
+  @override
+  Map<String, String> serialize() =>
+      super.serialize()..addAll(condition.serialize());
+
+  @override
+  String get kind => 'waitForCondition';
+}
+
+/// A Flutter Driver command that waits until there are no more transient callbacks in the queue.
+class WaitUntilNoTransientCallbacks extends Command {
+  /// Creates a command that waits for there to be no transient callbacks.
+  const WaitUntilNoTransientCallbacks({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilNoTransientCallbacks.deserialize(Map<String, String> json)
+      : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilNoTransientCallbacks';
+}
+
+/// A Flutter Driver command that waits until the frame is synced.
+class WaitUntilNoPendingFrame extends Command {
+  /// Creates a command that waits until there's no pending frame scheduled.
+  const WaitUntilNoPendingFrame({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilNoPendingFrame.deserialize(Map<String, String> json)
+      : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilNoPendingFrame';
+}
+
+/// A Flutter Driver command that waits until the Flutter engine rasterizes the
+/// first frame.
+///
+/// {@template flutter.frame_rasterized_vs_presented}
+/// Usually, the time that a frame is rasterized is very close to the time that
+/// it gets presented on the display. Specifically, rasterization is the last
+/// expensive phase of a frame that's still in Flutter's control.
+/// {@endtemplate}
+class WaitUntilFirstFrameRasterized extends Command {
+  /// Creates this command.
+  const WaitUntilFirstFrameRasterized({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilFirstFrameRasterized.deserialize(Map<String, String> json)
+      : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilFirstFrameRasterized';
 }
